First Working Version
A decent version of the clock that actually works and is somewhat
accurate. The acurassy clearly depends on the oscillator being used, in
this case I don't have any working crystals at this time. So an Arduino
is being used instead to generate the 32.768kHz signal used by timer1 to
keep track of the time of day.