First Working Version
A final working version of the clock. README.md shall be updated with
pinouts and a schematic of the entire clock, there are a few
configuration settings in the code that slightly modify the
functionality of the clock, including a speed adjustment to how fast the
numbers increment when setting the time.